Separate installTarget implementation from commands.go
